Add MediaOutputSwitcherDialogUI to TV startables.
This is a second entry point to the MediaOutputDialog introduced in U
that is used by the androidx utility method for opening it. The dialog
can be opened by any app using a broadcast either way so it doesn't
change what apps are able to do.

Bug: 290199629
Test: calling
androidx.mediarouter.mediarouter.SystemOutputSwitcherDialogController#showDialog(Context) opens the dialog

Change-Id: I3e7374672ba11c42b0406c681b3ea37fa22d8929