Exclude components from phone SystemUI in manifest
This CL adds a manifest file, whose purpose is to list all components
TVSystemUI is inheriting from SystemUI-core, which it does not
utilize. This enables R8 to no longer treat these components as entry
points, and thus apply code shrinking to them.

This change in isolation leads to ~750 KB of APK size savings.

Bug: 295161958
Test: make + flash
Change-Id: Idfd9192eec35e073fb50a76cd7722191655d72b6