Add CustomizationTestUtils as dep for SystemUI-tests-utils.
The frameworks/base CL in this topic adds a piece of code that requires
SystemUI-tests-utils to have access to SystemUICustomizationTestUtils.
Unfortunately SystemUI-tests-utils is a filegroup rather than a library,
so we need to add this dependency to every location that uses this
filegroup.

The ideal solution would be to have a SystemUITestUtils android_library
instead, but that's more complicated to achieve. See b/315806189 for a
related effort.

Bug: 293167744
Test: topic builds
Flag: N/A
Change-Id: Ia81c9654991b3ec86a5b16a8465ed522a9be94ad