Audio output: add a11y settings entry point
DPAD_RIGHT gets consumed by accessibility focus when Talkback is in use
and can therefore not be used to enter the settings.

Due to that the tooltip shouldn't show as it's not accurate. Showing a
button next to the output instead if settings are available.

Bug: 360215561
Test: manual - view output panel with and without talkback enabled,
check that outputs can still be selected, settings can be opened and no
views are clipped
Flag: com.android.settingslib.media.flags.enable_tv_media_output_dialog

Change-Id: Ic57dce66ecf8e17ddc4412f4c66730b7d098ef78