Remove mobile audio output switcher dependencies
Adds relevant functionality to the controller and creates TvMediaItems
as well since that was the only other remaining link.
Sorting and grouping logic is moved to the list proxy for easier
move over of the code and better readability.

Bug: 411440274
Test: open output switcher and check device list as well as switching
between the devices including pairing and unpairing bluetooth devices.
Flag: EXEMPT bugfix

Change-Id: Iceadcfb9cd13bb73beae592be95e107c7327f43e