Removed Deprecated Version of Shortcut Helper
KeyboardShortcuts is being removed, and replaced with a new version
"ShortcutHelper". See [ShortcutHelperCoreStartable].

This isn't enabled by default in TV as it wasn't required during the
refactor. However if the Keyboard Shortcuts Helper is need for TV,
please include ShortcutHelperModule in TVSystemUIModule and
ShortcutHelperDisplayModule in PerDisplayTvSystemUIModule.

Flag: EXEMPT FLAG_REMOVAL
Test: Build Test
Fix: 435405068
Change-Id: I5119657b27bb4ea5a4d4f7406e4b6c033a6e8e31